+++ extras/tests/Misc/Utf8.cpp
@@ -12,8 +12,8 @@ using namespace ARDUINOJSON_NAMESPACE;
 static void testCodepoint(uint32_t codepoint, std::string expected) {
   char buffer[4096];
   MemoryPool pool(buffer, 4096);
-  StringCopier str;
-  str.startString(&pool);
+  StringCopier str(pool);
+  str.startString();
 
   CAPTURE(codepoint);
   Utf8::encodeCodepoint(codepoint, str);